main.js
This update add the possibility to use "Blank.png" layers and hide them from the metadata.
This way  you can set this"Blank" layers to show up most of the times (Blank#90.png) and, just like that, some traits will not show up in every single NFT generated.
In example: you may have a trait called Special Hair with 5 different hair types in there. If you do not have a Blank file in that folder, as a layer, this kind of trait will always appear in the NFT generated. Of course each one will appear, accordingly with its rarity, but will show up in every single NFT created.
You can also work around this with a Grow Edition option, but is more straight forward using blank layers with higher chance of appearing.
The this is, if we use this blank layers, they will show up as Traits in Open Sea and we don't want that.
So... with the code added at line 176, basically it trims the layers names (so it doesn't have spaces before or after the text, and searches for the word "blank" while making it lowercase, in case you write it as uppercase). When it finds a layer called "blank" or "Blank" it will hide it from the metadata.
Just be careful a not use any layer as Blank if you want it to show up in your metadata.